# src/swagger.yaml
swagger: '2.0'
info:
  description: Tài liệu API cho Hệ thống quản lý học viên
  version: '1.0.0'
  title: Hệ thống quản ý Học viên
  contact:
    email: trinhthang1704@gmail.com
schemes:
  - http
  - https
security:
  - Bearer: []
basePath: /
paths:
...
  /users:
    patch:
      tags:
        - users
      summary: Sử dụng cho đổi mật khẩu
      x-swagger-router-controller: Users
      operationId: updatePassword
      parameters:
      - $ref: '#/parameters/UpdatePasswordParameter'
      produces:
        - application/json
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
  /users/student:
    post:
      tags:
        - users
      summary: Tạo một học viên 
      x-swagger-router-controller: Users
      operationId: createStudent 
      parameters:
      - $ref: '#/parameters/CreateStudentParameter'
      produces:
        - application/json
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
  /users/teacher:
    post:
      tags:
        - users
      summary: Tạo một giảng viên 
      x-swagger-router-controller: Users
      operationId: createTeacher
      parameters:
      - $ref: '#/parameters/CreateTeacherParameter'
      produces:
        - application/json
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
  /users/students:
    post:
      tags:
        - users
      summary: Tạo nhiều học viên  bằng  file xlsx
      x-swagger-router-controller: Users
      consumes:
        - multipart/form-data
      operationId: createStudents
      parameters:
      - $ref: '#/parameters/CreateStudentsParameter'
      produces:
        - application/json
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
  /users/teachers:
    post:
      tags:
        - users
      summary: Tạo nhiều giảng viên bằng file xlsx
      x-swagger-router-controller: Users
      operationId: createTeachers
      consumes:
        - multipart/form-data
      parameters:
      - $ref: '#/parameters/CreateTeachersParameter'
      produces:
        - application/json
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
  /users/{type}:
    get:
      tags:
        - users
      summary: Lấy danh sách học viên hoặc giảng viên
      x-swagger-router-controller: Users
      operationId: getUsers
      parameters:
      - $ref: '#/parameters/TypeParameter'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/UsersCollection'
        '400':
          description: Bad Request
...
parameters:
...
  UpdatePasswordParameter:
    in: body
    name: body
    required: true
    schema:
      $ref: "#/definitions/UpdatePasswordResponse"
  CreateStudentParameter:
    in: body
    name: body
    schema:
      $ref: "#/definitions/CreateStudentResponse"
  CreateTeacherParameter:
    in: body
    name: body
    schema:
      $ref: "#/definitions/CreateTeacherResponse"
  CreateStudentsParameter:
    in: formData
    name: file
    type: file
    description: file được tải lên.
  CreateTeachersParameter:
    in: formData
    name: file
    type: file
    description: file được tải lên..
  TypeParameter:
    in: path
    name: type
    required: true
    type: string
    enum: ['students', 'teachers']
definitions:
...
  UpdatePasswordResponse:
    type: object
    properties:
      username:
        type: string
      oldPassword:
        type: string
      newPassword:
        type: string
  CreateStudentResponse:
    type: object
    properties:
      msv: 
        type: string
      name:
        type: string
      address:
        type: string
      email:
        type: string
      phone:
        type: string
      schoolYearId:
        type: number
      classId:
        type: number
      instituaId:
        type: number
  CreateTeacherResponse:
    type: object
    properties:
      name:
        type: string
      address:
        type: string
      email:
        type: string
      phone:
        type: string
      instituaId:
        type: number
  UsersCollection:
    type: array
    items:
      type: object
      properties:
        id: 
          type: number
        msv: 
          type: string
        name: 
          type: string
        class: 
          type: string
        institua:
          type: string
        addres:
          type: string
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: 'Example: "Bearer ABC123"'